+++ attributes/default.rb
@@ -1,5 +1,5 @@
 # Latest RabbitMQ.com version to install
-default['rabbitmq']['version'] = '3.0.4'
+default['rabbitmq']['version'] = '3.1.5'
 # The distro versions may be more stable and have back-ported patches
 default['rabbitmq']['use_distro_version'] = false
 
@@ -26,6 +26,9 @@ default['rabbitmq']['default_pass'] = 'guest'
 # bind erlang networking to localhost
 default['rabbitmq']['local_erl_networking'] = false
 
+# bind rabbit and erlang networking to an address
+default['rabbitmq']['erl_networking_bind_address'] = nil
+
 #clustering
 default['rabbitmq']['cluster'] = false
 default['rabbitmq']['cluster_disk_nodes'] = []
@@ -34,6 +37,10 @@ default['rabbitmq']['erlang_cookie'] = 'AnyAlphaNumericStringWillDo'
 # resource usage
 default['rabbitmq']['disk_free_limit_relative'] = nil
 default['rabbitmq']['vm_memory_high_watermark'] = nil
+default['rabbitmq']['max_file_descriptors'] = 1024
+
+# job control
+default['rabbitmq']['job_control'] = 'initd'
 
 #ssl
 default['rabbitmq']['ssl'] = false
@@ -46,6 +53,7 @@ default['rabbitmq']['ssl_fail_if_no_peer_cert'] = false
 
 #virtualhosts
 default['rabbitmq']['virtualhosts'] = []
+default['rabbitmq']['disabled_virtualhosts'] = []
 
 #users
 default['rabbitmq']['enabled_users'] =
@@ -62,9 +70,8 @@ default['rabbitmq']['disabled_plugins'] = []
 case node['platform_family']
 when 'debian'
   default['rabbitmq']['package'] = "https://www.rabbitmq.com/releases/rabbitmq-server/v#{node['rabbitmq']['version']}/rabbitmq-server_#{node['rabbitmq']['version']}-1_all.deb"
-when 'rhel'
+when 'rhel','fedora'
   default['rabbitmq']['package'] = "https://www.rabbitmq.com/releases/rabbitmq-server/v#{node['rabbitmq']['version']}/rabbitmq-server-#{node['rabbitmq']['version']}-1.noarch.rpm"
-
 when 'smartos'
   default['rabbitmq']['service_name'] = 'rabbitmq'
   default['rabbitmq']['config_root'] = '/opt/local/etc/rabbitmq'
@@ -81,3 +88,4 @@ default['rabbitmq']['policies']['ha-two']['pattern'] = "^two\."
 default['rabbitmq']['policies']['ha-two']['params'] = { "ha-mode" => "exactly", "ha-params" => 2 }
 default['rabbitmq']['policies']['ha-two']['priority'] = 1
 
+default['rabbitmq']['disabled_policies'] = []