+++ recipes/virtualhost_management.rb
@@ -1,8 +1,10 @@
+# -*- coding: utf-8 -*-
 #
 # Cookbook Name:: rabbitmq
 # Recipe:: virtualhost_management
 #
 # Copyright 2013, Grégoire Seux
+# Copyright 2013, Opscode, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,12 +21,17 @@
 
 include_recipe "rabbitmq::default"
 
-virtualhosts = node['rabbitmq']['virtualhosts']
-service_name = node['rabbitmq']['service_name']
-virtualhosts.each do |virtualhost|
+node['rabbitmq']['virtualhosts'].each do |virtualhost|
   rabbitmq_vhost virtualhost do
     action :add
-    notifies :restart, "service[#{service_name}]"
+    notifies :restart, "service[#{node['rabbitmq']['service_name']}]"
+  end
+end
+
+node['rabbitmq']['disabled_virtualhosts'].each do |virtualhost|
+  rabbitmq_vhost virtualhost do
+    action :delete
+    notifies :restart, "service[#{node['rabbitmq']['service_name']}]"
   end
 end