+++ test/cookbooks/rabbitmq_test/files/default/tests/minitest/support/helpers.rb
@@ -1,5 +1,5 @@
 #
-# Copyright 2012, Opscode, Inc.
+# Copyright 2012-2013, Opscode, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,15 +16,30 @@
 
 module Helpers
   module RabbitMQ
-    require 'chef/mixin/shell_out'
-    include Chef::Mixin::ShellOut
+    require 'mixlib/shellout'
     include MiniTest::Chef::Assertions
     include MiniTest::Chef::Context
     include MiniTest::Chef::Resources
 
-    def plugin_enabled?(plugin_name)
-      plugin_list = shell_out("rabbitmq-plugins list -e '#{plugin_name}'")
-      plugin_list.stdout.include?(/\[[Ee]\] #{plugin_name}/)
+    def plugin_enabled?(plugin)
+      plugins = Mixlib::ShellOut.new("rabbitmq-plugins list -e '#{plugin}'").run_command
+      plugins.stdout =~ /(\[[Ee]\]\s#{plugin})/
     end
+
+    def policy_enabled?(policy)
+      policies = Mixlib::ShellOut.new("rabbitmqctl -q list_policies").run_command
+      policies.stdout =~ /\t#{policy}\t/
+    end
+
+    def user_enabled?(user)
+      users = Mixlib::ShellOut.new("rabbitmqctl -q list_users").run_command
+      users.stdout =~ /(#{user}\s)/
+    end
+
+    def vhost_enabled?(vhost)
+      vhosts = Mixlib::ShellOut.new("rabbitmqctl -q list_vhosts").run_command
+      vhosts.stdout =~ /(\n#{vhost}\n)/
+    end
+
   end
 end